Dummy data from database now live on dashboard page. Also big cleanup.
