+++ services/web-app/client/src/components/header-authenticated.js
@@ -15,33 +15,33 @@ function AccountNavButton() {
 }
 
 export default function HeaderAuthenticated() {
-    return (
-      <div className="App">
-        <Navbar expand="lg" style={{ backgroundColor: '#21212B' }}  data-bs-theme="dark">
-            <Container fluid>
-              <img src={securityseal} height={40} alt="Security Seal Logo"/>
-              <Navbar.Brand as={Link} to="/dashboard">Security Seal</Navbar.Brand>
-              <Navbar.Collapse id="basic-navbar-nav">
-                  <Nav className="me-auto">
-                    <Link to="/dashboard">
-                      <Button variant="none">
-                        Dashboard
-                      </Button>
-                    </Link>
-                    <Link to="/history">
-                      <Button variant="none">
-                        History
-                      </Button>
-                    </Link>
-                  </Nav>
-              </Navbar.Collapse>
-              <Navbar.Collapse className="justify-content-end">
-                <Navbar.Text>
-                      <AccountNavButton />
-                </Navbar.Text>
-              </Navbar.Collapse>
-            </Container>
-        </Navbar>
-      </div>
-    );
-  }
+  return (
+    <div className="App">
+      <Navbar expand="lg" style={{ backgroundColor: '#21212B' }}  data-bs-theme="dark">
+          <Container fluid>
+            <img src={securityseal} height={40} alt="Security Seal Logo"/>
+            <Navbar.Brand as={Link} to="/dashboard">Security Seal</Navbar.Brand>
+            <Navbar.Collapse id="basic-navbar-nav">
+                <Nav className="me-auto">
+                  <Link to="/dashboard">
+                    <Button variant="none">
+                      Dashboard
+                    </Button>
+                  </Link>
+                  <Link to="/history">
+                    <Button variant="none">
+                      History
+                    </Button>
+                  </Link>
+                </Nav>
+            </Navbar.Collapse>
+            <Navbar.Collapse className="justify-content-end">
+              <Navbar.Text>
+                    <AccountNavButton />
+              </Navbar.Text>
+            </Navbar.Collapse>
+          </Container>
+      </Navbar>
+    </div>
+  );
+}
