+++ services/web-app/client/src/components/vulnerability-cards.js
@@ -3,17 +3,18 @@ import Card from 'react-bootstrap/Card';
 import CodeBlock from './display-code';
 import 'bootstrap/dist/css/bootstrap.min.css';
 
-function VulnerabilityCards({ data }) { // Receive the data array directly as a prop
+
+function VulnerabilityCards({ vulnerabilities }) { // Expect an array of vulnerabilities
   return (
     <div>
-      {data.map((item, index) => ( // Map through the data array
-        <Card key={index} style={{ backgroundColor: '#21212B', color: '#FFFFFF', padding: 20 }}> {/* Use the index as a key for each card */}
-          <Card.Header>{item.CWE}</Card.Header>
+      {vulnerabilities.map((vuln, index) => (
+        <Card key={index} style={{ backgroundColor: '#21212B', color: '#FFFFFF', padding: 20 }}>
+          <Card.Header>{vuln.cwe.cwe_id}: {vuln.cwe.cwe_name}</Card.Header> 
           <Card.Body>
             <Card.Title>Vulnerability Detail</Card.Title>
             <div>
-              {item.info}
-              <CodeBlock codeString={item.code} />
+              {vuln.vuln_summary} 
+              <CodeBlock codeString={vuln.code_extract} /> 
             </div>
           </Card.Body>
         </Card>
