+++ services/web-app/client/src/pages/not-found-page.js
@@ -4,4 +4,4 @@ export default function NotFoundPage() {
         <h1 >404 - Page Not Found</h1>
         </div>
     );
-    }
+}
